projects/gta-update-alerts/gta-update-alerts-flask/requirements.txt
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==2.3.3
google-recaptcha==2.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Markdown==3.6
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pytz==2024.1
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1
Werkzeug==3.0.1
